move create_and_send_schema, create_and_send_credential_definition implementation to base class, fixes indentation error in indy_vdr method; cleanups
